+++ example/package.json
@@ -1,10 +1,10 @@
 {
     "name": "botbuilder-mongo-storage-sample",
     "dependencies": {
-        "botbuilder": "^4.2.0",
+        "botbuilder": "^4.8.0",
         "botbuilder-greeting": "^1.0.1",
-        "mongodb": "^3.1.10",
-        "restify": "^7.3.0"
+        "mongodb": "^3.5.6",
+        "restify": "^8.5.1"
     },
     "scripts": {
         "start": " node app.js"
