+++ src/MongoDbStorageConfig.ts
@@ -1,0 +1,7 @@
+import { MongoClientOptions } from 'mongodb';
+export interface MongoDbStorageConfig {
+  url: string;
+  database?: string;
+  collection?: string;
+  clientOptions?: MongoClientOptions;
+}
