+++ src/MongoDocumentStoreItem.ts
@@ -1,0 +1,4 @@
+export interface MongoDocumentStoreItem {
+  _id: string;
+  state: any;
+}
